Refactor a bit, clean up some code
- Add declaration.d.ts for css?inline
- Split AgGridView into AgGridWidget and AgGridView
  - AgGridWidget just loads the table from the widget, and then passes it to AgGridView
  - AgGridView just translates from a Table to what an AgGridReact needs
