+++ plugins/ag-grid/src/js/src/AgGridPlugin.ts
@@ -1,13 +1,13 @@
 import { type WidgetPlugin, PluginType } from '@deephaven/plugin';
 import { vsTable } from '@deephaven/icons';
 import type { Widget } from '@deephaven/jsapi-types';
-import AgGridView from './AgGridView';
+import { AgGridWidget } from './AgGridWidget';
 
 export const AgGridPlugin: WidgetPlugin<Widget> = {
   name: '@deephaven/js-plugin-ag-grid',
   type: PluginType.WIDGET_PLUGIN,
   supportedTypes: 'deephaven.aggrid.AgGrid',
-  component: AgGridView,
+  component: AgGridWidget,
   icon: vsTable,
 };
