+++ plugins/ag-grid/src/js/src/AgGridView.tsx
@@ -1,50 +1,24 @@
-import React, { useEffect, useMemo, useState } from 'react';
+import React, { useMemo } from 'react';
 import { ColDef, IServerSideDatasource } from '@ag-grid-community/core';
 import styles from '@ag-grid-community/styles/ag-grid.css?inline'; // Core CSS
 import quartzStyles from '@ag-grid-community/styles/ag-theme-quartz.css?inline'; // Theme
 import { AgGridReact } from '@ag-grid-community/react';
 import { ServerSideRowModelModule } from '@ag-grid-enterprise/server-side-row-model';
-import type { Table, Widget } from '@deephaven/jsapi-types';
-import { type WidgetComponentProps } from '@deephaven/plugin';
-import { useApi } from '@deephaven/jsapi-bootstrap';
+import type { Table } from '@deephaven/jsapi-types';
 import Log from '@deephaven/log';
 
 const log = Log.module('@deephaven/js-plugin-ag-grid/AgGridView');
 
+type AgGridViewProps = { table: Table };
+
 /**
- * Basic AgGrid view that loads a Deephaven table and displays it in AG Grid.
+ * Basic AgGrid view that uses a Deephaven table a data source and displays it in AG Grid.
  *
  * Does not support value formatting, sorting, filtering, or basically any functionality beyond
  * just displaying the columns and rows of data.
  */
-export function AgGridView(
-  props: WidgetComponentProps<Widget>
-): JSX.Element | null {
-  const dh = useApi();
-  const { fetch } = props;
-  const [table, setTable] = useState<Table>();
-
-  /** First we load the widget object. This is the object that is sent from the server in AgGridMessageStream. */
-  useEffect(() => {
-    let cancelled = false;
-    async function init() {
-      log.debug('Fetching widget');
-      const widget: Widget = await fetch();
-      const newTable = (await widget.exportedObjects[0].fetch()) as Table;
-      if (!cancelled) {
-        log.info('AgGridView loaded table', newTable);
-        setTable(newTable);
-      }
-    }
-
-    init();
-
-    return () => {
-      cancelled = true;
-    };
-  }, [dh, fetch]);
-
-  log.debug('AgGridView rendering', table, table?.columns);
+export function AgGridView({ table }: AgGridViewProps): JSX.Element | null {
+  log.info('AgGridView rendering', table, table?.columns);
 
   /** Map from Deephaven Table Columns to AG Grid ColDefs */
   const colDefs: ColDef[] = useMemo(
