+++ plugins/ag-grid/src/js/src/declaration.d.ts
@@ -1,0 +1,6 @@
+declare module '*.css?inline' {
+  const content: string;
+  export default content;
+}
+
+declare module '*.scss';
